Remove actual search functionality
Seems unnecessary really. Users of this library may want to do all kinds
of different things with the results of their searches and providing a
search function to try and do all of that is a bad idea. Now we just have
a bunch of helpers to do with searching.
