--- tests/search/__init__.py
@@ -1,2 +1,0 @@
-#!/usr/bin/env python
-# encoding: utf-8
