--- tests/search/test_search.py
@@ -1,2 +1,0 @@
-#!/usr/bin/env python
-# encoding: utf-8
